requirements.txt
farm-haystack[all-gpu]
streamlit
st-annotated-text
